Rename appfeaturelistcallback to featurelistcallback
