+++ src/cmake/sources.cmake
@@ -27,7 +27,7 @@ set_property(TARGET mozillavpn-sources PROPERTY INTERFACE_INCLUDE_DIRECTORIES
 target_sources(mozillavpn-sources INTERFACE
     ${CMAKE_CURRENT_SOURCE_DIR}/constants.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/constants.h
-    ${CMAKE_CURRENT_SOURCE_DIR}/appfeaturelistcallback.h
+    ${CMAKE_CURRENT_SOURCE_DIR}/featurelistcallback.h
     ${CMAKE_CURRENT_SOURCE_DIR}/appfeaturelist.h
     ${CMAKE_CURRENT_SOURCE_DIR}/appimageprovider.h
     ${CMAKE_CURRENT_SOURCE_DIR}/applistprovider.h
