+++ tests/auth_tests/CMakeLists.txt
@@ -66,7 +66,6 @@ target_compile_definitions(app_auth_tests PRIVATE MZ_DUMMY)
 # Unit test source files
 target_sources(app_auth_tests PRIVATE
     appconstants.h
-    appfeaturelistcallback.h
     appfeaturelist.h
     appsettingslist.h
     main.cpp
